requirements.txt
FlaskApp
requests